requirements.txt
pinecone-client
python-dotenv
pymupdf
transformers
torch